Increase the cruz hitbox
The cruz actually had a hitbox as large as any other enemy. Yet
it did feel like it was harder to hit. I think it was an illusion caused
by it's arms. I extended the hit box to just before the length of the
arms. Extending the hitbox to the length of the arms, I think would make
the hitbox far too large.