Try to increase the fall rate of the hearts
What I did was change their spawn "force" to a spawn velocity instead.
This way the speed at which they spawn is independant of their mass. As
for the fall rate, mass does not affect the downward acceleration, but I
can control the initial upward velocity. So I set it to 0 so that the
heart begins falling as soon as it spawns. Hopefully, this fall rate
is fast enough to be desirable.